Fix issue with emojis before points
If a user is awarded an emoji but has no points their score will be
displayed as NaN in the scores table.

To prevent this, check the user's score when they are awarded an emoji.
If they do not yet have a score an rejected promise will be returned -
catch this and set their score to 0.